move tooltip logic for string y-axis tick labels to cartesian chart
