+++ packages/charts/react-charting/src/components/HeatMapChart/HeatMapChart.types.ts
@@ -118,6 +118,12 @@ export interface IHeatMapChartProps extends Pick<ICartesianChartProps, Exclude<k
    * @default alphabetical
    * The prop used to decide order of string axis labels */
   sortOrder?: 'none' | 'alphabetical';
+
+  /**
+   *@default false
+   *Used for to elipse y axis labes and show tooltip on x axis labels
+   */
+  showYAxisLablesTooltip?: boolean;
 }
 
 /**
