+++ packages/charts/react-charting/src/components/HorizontalBarChartWithAxis/HorizontalBarChartWithAxis.styles.ts
@@ -1,25 +1,12 @@
 import { IHorizontalBarChartWithAxisStyleProps, IHorizontalBarChartWithAxisStyles } from '../../index';
 
 export const getStyles = (props: IHorizontalBarChartWithAxisStyleProps): IHorizontalBarChartWithAxisStyles => {
-  const { shouldHighlight, theme } = props;
+  const { shouldHighlight } = props;
   return {
     opacityChangeOnHover: {
       opacity: shouldHighlight ? '' : '0.1',
     },
 
     xAxisTicks: [],
-
-    tooltip: {
-      ...theme.fonts.medium,
-      display: 'flex',
-      flexDirection: 'column',
-      padding: '8px',
-      position: 'absolute',
-      textAlign: 'center',
-      top: '0px',
-      background: theme.semanticColors.bodyBackground,
-      borderRadius: '2px',
-      pointerEvents: 'none',
-    },
   };
 };
